// lab_web/src/schemas/gen/coco/annotation/v1/annotation_pb.ts
// Copyright 2026 Gautham Anil
// SPDX-License-Identifier: Apache-2.0

// Text attached to a moment (and optionally a place): captions, notes,
// "replanning because the door closed".
// Channel: coco.annotation.text.v1 (AnnotationBatch).

// @generated by protoc-gen-es v2.15.0 with parameter "target=ts,import_extension=none"
// @generated from file coco/annotation/v1/annotation.proto (package coco.annotation.v1, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file coco/annotation/v1/annotation.proto.
 */
export const file_coco_annotation_v1_annotation: GenFile = /*@__PURE__*/
  fileDesc("CiNjb2NvL2Fubm90YXRpb24vdjEvYW5ub3RhdGlvbi5wcm90bxISY29jby5hbm5vdGF0aW9uLnYxIrEBCg9Bbm5vdGF0aW9uQmF0Y2gSCwoDc2VxGAEgAygEEgwKBHRpY2sYAiADKAQSDwoHdF93b3JsZBgDIAMoARIMCgR0ZXh0GAQgAygJEigKBWxldmVsGAUgAygOMhkuY29jby5hbm5vdGF0aW9uLnYxLkxldmVsEhQKDGhhc19wb3NpdGlvbhgGIAMoCBIJCgF4GAcgAygBEgkKAXkYCCADKAESDgoGc291cmNlGAkgAygJKlMKBUxldmVsEhUKEUxFVkVMX1VOU1BFQ0lGSUVEEAASDgoKTEVWRUxfSU5GTxABEhAKDExFVkVMX05PVElDRRACEhEKDUxFVkVMX1dBUk5JTkcQA2IGcHJvdG8z");

/**
 * @generated from message coco.annotation.v1.AnnotationBatch
 */
export type AnnotationBatch = Message<"coco.annotation.v1.AnnotationBatch"> & {
  /**
   * @generated from field: repeated uint64 seq = 1;
   */
  seq: bigint[];

  /**
   * @generated from field: repeated uint64 tick = 2;
   */
  tick: bigint[];

  /**
   * @generated from field: repeated double t_world = 3;
   */
  tWorld: number[];

  /**
   * @generated from field: repeated string text = 4;
   */
  text: string[];

  /**
   * @generated from field: repeated coco.annotation.v1.Level level = 5;
   */
  level: Level[];

  /**
   * @generated from field: repeated bool has_position = 6;
   */
  hasPosition: boolean[];

  /**
   * @generated from field: repeated double x = 7;
   */
  x: number[];

  /**
   * @generated from field: repeated double y = 8;
   */
  y: number[];

  /**
   * who wrote it: "coco_lab", "lens", "author"
   *
   * @generated from field: repeated string source = 9;
   */
  source: string[];
};

/**
 * Describes the message coco.annotation.v1.AnnotationBatch.
 * Use `create(AnnotationBatchSchema)` to create a new message.
 */
export const AnnotationBatchSchema: GenMessage<AnnotationBatch> = /*@__PURE__*/
  messageDesc(file_coco_annotation_v1_annotation, 0);

/**
 * @generated from enum coco.annotation.v1.Level
 */
export enum Level {
  /**
   * @generated from enum value: LEVEL_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: LEVEL_INFO = 1;
   */
  INFO = 1,

  /**
   * @generated from enum value: LEVEL_NOTICE = 2;
   */
  NOTICE = 2,

  /**
   * @generated from enum value: LEVEL_WARNING = 3;
   */
  WARNING = 3,
}

/**
 * Describes the enum coco.annotation.v1.Level.
 */
export const LevelSchema: GenEnum<Level> = /*@__PURE__*/
  enumDesc(file_coco_annotation_v1_annotation, 0);